requirements.txt
# COCOA Requirements
# Consciousness Orchestration and Cognitive Operations

# Core Dependencies
python-dotenv>=1.0.0

# Database & Memory
numpy>=1.24.0

# AI & Language Processing
anthropic>=0.64.0
openai>=1.0.0
tavily-python>=0.7.0

# Terminal UI
rich>=13.7.0
prompt-toolkit>=3.0.0

# Audio System (installed via setup_audio.sh)
elevenlabs>=2.11.0
pygame>=2.1.0
scipy>=1.9.0
soundfile>=0.12.0